requirements.txt
pytest==7.4.2
pytest-django==4.5.2